penumbra adjustments

+++ Dresser/Services/PenumbraIpc.cs
@@ -21,7 +21,7 @@ using System.Threading.Tasks;
 using static Dresser.Services.Storage;
 
 using CurrentSettings = System.ValueTuple<Penumbra.Api.Enums.PenumbraApiEc, (bool EnabledState, int Priority, System.Collections.Generic.IDictionary<string, System.Collections.Generic.IList<string>> EnabledOptions, bool Inherited)?>;
-using PseudoEquipItem = System.ValueTuple<string, ulong, ushort, ushort, ushort, byte, uint>;
+using PseudoEquipItem = System.ValueTuple<string, ulong, uint, ushort, ushort, byte, uint>;
 
 namespace Dresser.Services;
